[feat] add gates' explanations
- add gates' explanations
- instantiate paper items when doors open
