+++ Igloo/Assets/Animations/LogicGates.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: ea5fca43fe080074eb3cb7bac2d07823
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
