+++ Igloo/Assets/Animations/LogicGates/ANDGate.controller.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: b21a0f4ddd33ce94f89ed98ecd23c321
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 9100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
