+++ Igloo/Assets/Animations/LogicGates/Floating.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: b18928a812514a24cb76b31834927fc3
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 7400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
