+++ Igloo/Assets/Animations/LogicGates/FloatingAND.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 2cd7c538a1089a541b9338f9aa6e0597
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 7400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
