+++ Igloo/Assets/Animations/LogicGates/FloatingOR.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 8960a95df12403b49923ee005304f46d
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 7400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
