+++ Igloo/Assets/Animations/LogicGates/ORGate.controller.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 43b10ff90054dca46b0c2574c93c7f40
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 9100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
