+++ Igloo/Assets/Animations/LogicGates/notGate.controller.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: ca490a706304ca746b982c232d3e2319
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 9100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
