+++ Igloo/Assets/Code/Scripts/Dungeon/DigitalDesign/FifthDoor.cs
@@ -4,10 +4,27 @@ using UnityEngine;
 
 public class FifthDoor : Door
 {
+    public GameObject itemPrefab;
+    private bool isFirst = true;
+
     void Start()
     {
         levers = new bool[4];
         animator = GetComponent<Animator>();
+    }
+
+    void Update()
+    {
+        if (isOpen)
+        {
+            animator.SetTrigger("open");
+
+            if (isFirst)
+            {
+                GameObject instance = Instantiate(itemPrefab);
+                isFirst = false;
+            }
+        }
     }
 
     public override void CheckLevers()
