+++ Igloo/Assets/Code/Scripts/Interactables/DigitalDesign.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 53e3a341582fa444fbc1c736e563d324
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
