+++ Igloo/Assets/Code/Scripts/Interactables/DigitalDesign/ExplanationAND.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: d8d1432924411c54ab341a6e7d8149cb
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
