+++ Igloo/Assets/Code/Scripts/Interactables/DigitalDesign/ExplanationNOT.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 1beda6d4c2fa2d24ba294927f705df77
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
