+++ Igloo/Assets/Code/Scripts/Interactables/DigitalDesign/ExplanationOR.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: b6565f1d83ae43640952abe3fe1dcc29
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
