--- Igloo/Assets/Code/Scripts/Interactables/TestLever.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: f7d84caf06d730f4f907c7b35e13a60b
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
