--- Igloo/Assets/Code/Scripts/Interactables/TrueLever.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 5bea78e044c51f04ca63ab3ebb8002b7
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
