+++ Igloo/Assets/Prefabs/Dungeon/DigitalDesign/PaperItem1.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: ef6946fdfeca972418a43c426c2c7e3a
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
