+++ Igloo/Assets/Prefabs/Dungeon/DigitalDesign/PaperItem2.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 4e75810a487dad54eafdc0153d857690
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
