+++ Igloo/Assets/Prefabs/Dungeon/DigitalDesign/notGate.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 2c6071d9b80d83940837a41af5f930ca
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
